requirements.txt
django
djangorestframework
markdown
django-filter
gensim
nltk
xlrd
keras
tensorflow
